Working terrain data message

+++ Assets/Scripts/TerrainScripts/Details/TerrainResourceNode.cs
@@ -7,6 +7,6 @@ namespace Assets.Scripts.TerrainScripts.Details
     public class TerrainResourceNode
     {
         public ResourcePrefabsList prefabsList;
-        public int resourceTypeID;
+        public byte resourceTypeID;
     }
 }
